backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
grpcio>=1.64.1
grpcio-tools>=1.64.1
redis>=5.0.4
python-dotenv>=1.0.1
PyJWT>=2.8.0
